move helpers to service folder, make toasts position fixed while scrolling
