+++ src/app/intro/intro.component.tns.html
@@ -11,6 +11,7 @@
     <Button kirby-button text="Grid" [nsRouterLink]="['/examples/grid']" class="btn btn-primary"></Button>
     <Button kirby-button text="Action Sheet" [nsRouterLink]="['/examples/action-sheet']"
       class="btn btn-primary"></Button>
+    <Button kirby-button text="Toast" [nsRouterLink]="['/examples/toast']" class="btn btn-primary"></Button>
     <Button kirby-button text="Alert" [nsRouterLink]="['/examples/alert']" class="btn btn-primary"></Button>
     <Button kirby-button text="Modal" [nsRouterLink]="['/examples/modal']" class="btn btn-primary"></Button>
     <Button kirby-button text="Fonts" [nsRouterLink]="['/examples/fonts']"></Button>
