+++ src/kirby/components/toast/services/toast.helper.tns.ts
@@ -1,0 +1,10 @@
+import { Injectable } from '@angular/core';
+
+import { ToastConfig } from '../config/toast-config';
+
+@Injectable()
+export class ToastHelper {
+  public async showToast(config: ToastConfig): Promise<any> {
+    return Promise.resolve();
+  }
+}
